D3 Exploding Boxplots

Produce an svg element showing boxplots that explode upon mouse-click into jittered individual-value plots.

bpexploder(data, settings = list(groupVar = NULL, levels = NULL, levelLabels =
  NULL, levelColors = NULL, yVar, yAxisLabel = NULL, xAxisLabel = NULL, tipText
  = NULL, referenceId = NULL, relativeWidth = 1), width = NULL,
  height = NULL, elementId = NULL)

Arguments

data
A data frame containing all variables involved in the plot.

settings
A list with the following elements:

groupVar Character vector of length one setting name of grouping variable. (Leave NULL or omit for no grouping.)

levels Character vector of values of groupVar. Leave NULL or omit to render levels in alphabetical order.

levelLabels Character vector setting custom labels for levels of grouping variable.

levelColors Character vector setting custom colors for boxes.

yVar Character vector of length one, specifying the numerical variable to plot.

yAxisLabel Character vector of length one.

xAxisLabel Character vector of length one.

tipText List of character vectors. Names are variables to show in the tooltips, values are the labels to denote these variables.

referenceId Character vector; id attribute of an HTML tag. Offset-width of this element is used to set width of the svg. Defaults to the grandparent element.

relativeWidth Number between 0 and 1. The svg-width is this number multplied by the offset-width of the reference element.

width, height
Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended. Suggest to leave NULL.

elementId
id of element ot contain widget. Should be left NULL.